requirements.txt
# DataMosaic 项目 Python 依赖包
# 按照功能分类列出所有必需的包

# ======== 核心Web框架 ========
flask==2.3.3
flask-cors==4.0.0
werkzeug==2.3.7

# ======== LangChain相关 ========
langchain==0.1.20
langchain-community==0.0.38

# ======== 文档处理 ========
pypdf2==3.0.1
unstructured==0.10.30
python-docx==0.8.11
openpyxl==3.1.2
pandas==2.0.3

# ======== AI/ML模型相关 ========
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
numpy==1.24.3

# ======== HTTP和环境变量 ========
requests==2.31.0
python-dotenv==1.0.0

